fix(VectorStore): Add expire value to VectorStore - Add an expiration value to the `VectorStore` to manage the lifecycle of the stored vectors. - and a little bit code design refactoring
