+++ src/WK.OpenAiWrapper/Client.storage..cs
@@ -1,6 +1,5 @@
 ﻿using WK.OpenAiWrapper.Result;
 using WK.OpenAiWrapper.Models;
-using WK.OpenAiWrapper.Interfaces;
 using WK.OpenAiWrapper.Services;
 using OpenAI.VectorStores;
 
@@ -8,31 +7,32 @@ namespace WK.OpenAiWrapper;
 
 internal partial class Client
 {
-    internal readonly FileService FileService;
+    internal readonly StorageService StorageService;
+    
     public async Task<Result<OpenAiMultipleFilesVectorStoreResponse>> UploadToNewVectorStore(string[] filePaths, string vectorStoreName, bool waitForDoneStatus = false) 
-        => await FileService.UploadToNewVectorStore(filePaths, vectorStoreName, waitForDoneStatus).ConfigureAwait(false);
+        => await StorageService.UploadToNewVectorStore(filePaths, vectorStoreName, waitForDoneStatus).ConfigureAwait(false);
 
     public async Task<Result<OpenAiMultipleFilesVectorStoreResponse>> UploadToVectorStore(string[] filePaths, string vectorStoreId, bool waitForDoneStatus = false) 
-        => await FileService.UploadToVectorStore(filePaths, vectorStoreId, waitForDoneStatus).ConfigureAwait(false);
+        => await StorageService.UploadToVectorStore(filePaths, vectorStoreId, waitForDoneStatus).ConfigureAwait(false);
 
     public async Task<Result<OpenAiVectorStoreResponse>> UploadStreamToNewVectorStore(Stream fileStream, string fileName, string vectorStoreName, bool waitForDoneStatus = false) =>
-        await FileService.UploadStreamToNewVectorStore(fileStream, fileName, vectorStoreName, waitForDoneStatus).ConfigureAwait(false);
+        await StorageService.UploadStreamToNewVectorStore(fileStream, fileName, vectorStoreName, waitForDoneStatus).ConfigureAwait(false);
 
     public async Task<Result<OpenAiMultipleFilesVectorStoreResponse>> UploadToNewVectorStore(List<(string FileName, byte[] FileBytes)> files, string vectorStoreName, bool waitForDoneStatus = false) 
-        => await FileService.UploadToNewVectorStore(files, vectorStoreName, waitForDoneStatus).ConfigureAwait(false);
+        => await StorageService.UploadToNewVectorStore(files, vectorStoreName, waitForDoneStatus).ConfigureAwait(false);
 
     public async Task<Result<OpenAiMultipleFilesVectorStoreResponse>> UploadToVectorStore(List<(string FileName, byte[] FileBytes)> files, string vectorStoreId, bool waitForDoneStatus = false) 
-        => await FileService.UploadToVectorStore(files, vectorStoreId, waitForDoneStatus).ConfigureAwait(false);
+        => await StorageService.UploadToVectorStore(files, vectorStoreId, waitForDoneStatus).ConfigureAwait(false);
 
     public async Task<Result<OpenAiVectorStoreResponse>> DeleteFileInVectorStoreByName(string fileName, string vectorStoreId) 
-        => await FileService.DeleteFileInVectorStoreByName(fileName, vectorStoreId).ConfigureAwait(false);
+        => await StorageService.DeleteFileInVectorStoreByName(fileName, vectorStoreId).ConfigureAwait(false);
 
     public async Task<Result<OpenAiVectorStoreResponse>> DeleteFileInVectorStoreById(string fileId, string vectorStoreId) 
-        => await FileService.DeleteFileInVectorStoreById(fileId, vectorStoreId).ConfigureAwait(false);
+        => await StorageService.DeleteFileInVectorStoreById(fileId, vectorStoreId).ConfigureAwait(false);
 
     public async Task<VectorStoreFileResponse> GetVectorStoreFileStatusAsync(string vectorStoreId, string fileId) 
-        => await FileService.GetVectorStoreFileStatusAsync(vectorStoreId, fileId).ConfigureAwait(false);
+        => await StorageService.GetVectorStoreFileStatusAsync(vectorStoreId, fileId).ConfigureAwait(false);
 
     public async Task<Result<OpenAiVectorStoreResponse>> UploadStreamToVectorStore(Stream fileStream, string fileName, string vectorStoreId, bool waitForDoneStatus = false) =>
-        await FileService.UploadStreamToVectorStore(fileStream, fileName, vectorStoreId, waitForDoneStatus).ConfigureAwait(false);
+        await StorageService.UploadStreamToVectorStore(fileStream, fileName, vectorStoreId, waitForDoneStatus).ConfigureAwait(false);
 }
